refactor: unlike product function
